requirements.txt
fastapi
uvicorn
openai
pydantic
groq
python-dotenv